+++ Core/libft/ft_putendl_fd.c
@@ -6,7 +6,7 @@
 /*   By: lfrench <lfrench@student.42luxembourg      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/11 14:39:27 by lfrench           #+#    #+#             */
-/*   Updated: 2024/03/02 20:26:39 by lfrench          ###   ########.fr       */
+/*   Updated: 2024/03/02 21:18:51 by lfrench          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 
 void	ft_putendl_fd(char *s, int fd)
 {
-	int		i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != 0)
